sql: make sql checks on server side
To make SQL CHECKS on space insertion in LUA we have to
generate a temporal VDBE and execute it with a tuple
mapped in VDBE memory with new opcode OP_FieldMapMemory as
an argument. These actions are performed by a new special trigger
set for the space on server side.
The CHECKS are carried out twice for SQL requests because of
context and affinity changes to which the inserting tuple is
subjected.
At first, user could specify IGNORE keyword for SQL UPDATE that
is a part of PARSER CONTEXT and couldn't be handled on
server-side for now.
The other point is that tuple represented as msgpack on
server-side may have already changed field data type.

E.g.
CREATE TABLE t2(id primary key,
                z TEXT CONSTRAINT three
                  CHECK(typeof(coalesce(z,'')) == 'text'));
INSERT INTO t2 VALUES(1, 3.14159);

Before the tuple for insertion into the server-side space was
generated, it's z field was of type REAL, which did not pass
the CHECK test. But then new compatible affinity BLOB is applied,
so making check on server-side no problems would be triggered
and tuple would be inserted successfully.

Closes #3691.