box: move info_handler interface into src/info
Box/info.h defines info_handler interface with a set
of virtual functions. It allows to hide Lua from code
not depending on this language, and is used in things
line index:info(), box.info() to build Lua table with
some info. But is does not depend on box/ so move it
to src/.

Also, this API is needed for the forthcoming SWIM
module which is going to be placed into src/.

Needed for #3234